Alter cluster-access-request utilities to assume that the AllocationUser already exists
